requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
django-rest-passwordreset
drf-yasg
redis
celery